Add helper function `fileInitializationFailed`
